+++ library/src/methods/fallback/fallbackAsync.ts
@@ -1,4 +1,4 @@
-import type { BaseSchema, BaseSchemaAsync, Output } from '../../types.ts';
+import type { BaseSchemaAsync, Output } from '../../types.ts';
 import { getOutput } from '../../utils/index.ts';
 import type { FallbackInfo } from './types.ts';
 
@@ -10,12 +10,28 @@ import type { FallbackInfo } from './types.ts';
  *
  * @returns The passed schema.
  */
-export function fallbackAsync<TSchema extends BaseSchema | BaseSchemaAsync>(
+export function fallbackAsync<
+  TSchema extends BaseSchemaAsync,
+  TFallback extends Output<TSchema>
+>(
   schema: TSchema,
-  value: Output<TSchema> | ((info: FallbackInfo) => Output<TSchema>)
-): TSchema {
+  fallback_:
+    | TFallback
+    | ((info?: FallbackInfo) => TFallback | Promise<TFallback>)
+): TSchema & { getFallback: (info?: FallbackInfo) => Promise<TFallback> } {
   return {
     ...schema,
+
+    /**
+     * Returns the default value.
+     */
+    async getFallback(info) {
+      return typeof fallback_ === 'function'
+        ? await (
+            fallback_ as (info?: FallbackInfo) => TFallback | Promise<TFallback>
+          )(info)
+        : (fallback_ as TFallback);
+    },
 
     /**
      * Parses unknown input based on its schema.
@@ -29,12 +45,7 @@ export function fallbackAsync<TSchema extends BaseSchema | BaseSchemaAsync>(
       const result = await schema._parse(input, info);
       return getOutput(
         result.issues
-          ? typeof value === 'function'
-            ? (value as (info: FallbackInfo) => Output<TSchema>)({
-                input,
-                issues: result.issues,
-              })
-            : value
+          ? await this.getFallback({ input, issues: result.issues })
           : result.output
       );
     },