+++ library/CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to the library will be documented in this file.
 - Add `getRestAndDefaultArgs` utility function
 - Add new `rest` argument to `object` and `objectAsync` schema
 - Add `discriminatedUnion` and `discriminatedUnionAsync` schema (issue #90, #216)
+- Add `getFallback` property to schema in `fallback` method (pull request #177)
 - Add `PartialObjectEntries` and `PartialObjectEntriesAsync` type (issue #217)
 - Add export for any validation regex (pull request #219)
 - Fix type check in `date` and `dateAsync` for invalid dates (pull request #214)